refactor(transfer): extract inline sections into page component folder
