+++ src/components/pages/transfer/types.ts
@@ -1,0 +1,21 @@
+export type FormErrors = {
+  recipient?: string
+  amount?: string
+  targetTick?: string
+}
+
+export type TxResult = {
+  txId: string
+  targetTick: string
+  amount: bigint
+  tokenName: string
+  sourceIdentity: string
+  recipient: string
+  fee: bigint
+}
+
+export type SourceAccount = {
+  name: string
+  identity: string
+  watchOnly?: boolean
+}
